+++ lib/init-action.js
@@ -198,7 +198,7 @@ async function run() {
     catch (unwrappedError) {
         const error = (0, util_1.wrapError)(unwrappedError);
         core.setFailed(error.message);
-        const statusReportBase = await (0, status_report_1.createStatusReportBase)(status_report_1.ActionName.Init, error instanceof util_1.ConfigurationError ? "user-error" : "aborted", startedAt, config, await (0, util_1.checkDiskUsage)(), logger, error.message, error.stack);
+        const statusReportBase = await (0, status_report_1.createStatusReportBase)(status_report_1.ActionName.Init, error instanceof util_1.ConfigurationError ? "user-error" : "aborted", startedAt, config, await (0, util_1.checkDiskUsage)(logger), logger, error.message, error.stack);
         if (statusReportBase !== undefined) {
             await (0, status_report_1.sendStatusReport)(statusReportBase);
         }
@@ -329,7 +329,7 @@ async function run() {
         if (!(await (0, util_1.codeQlVersionAtLeast)(codeql, "2.15.1")) &&
             process.platform === "darwin" &&
             (process.arch === "arm" || process.arch === "arm64") &&
-            !(await (0, init_1.isSipEnabled)(logger))) {
+            !(await (0, util_1.isSipEnabled)(logger))) {
             logger.warning("CodeQL versions 2.15.0 and lower are not supported on MacOS ARM machines with System Integrity Protection (SIP) disabled.");
         }
         // From 2.16.0 the default for the python extractor is to not perform any
