+++ lib/init.js
@@ -28,11 +28,9 @@ exports.initConfig = initConfig;
 exports.runInit = runInit;
 exports.printPathFiltersWarning = printPathFiltersWarning;
 exports.checkInstallPython311 = checkInstallPython311;
-exports.isSipEnabled = isSipEnabled;
 exports.cleanupDatabaseClusterDirectory = cleanupDatabaseClusterDirectory;
 const fs = __importStar(require("fs"));
 const path = __importStar(require("path"));
-const exec = __importStar(require("@actions/exec/lib/exec"));
 const toolrunner = __importStar(require("@actions/exec/lib/toolrunner"));
 const safeWhich = __importStar(require("@chrisgavin/safe-which"));
 const actions_util_1 = require("./actions-util");
@@ -91,26 +89,6 @@ async function checkInstallPython311(languages, codeql) {
         await new toolrunner.ToolRunner(await safeWhich.safeWhich("powershell"), [
             script,
         ]).exec();
-    }
-}
-// For MacOS runners: runs `csrutil status` to determine whether System
-// Integrity Protection is enabled.
-async function isSipEnabled(logger) {
-    try {
-        const sipStatusOutput = await exec.getExecOutput("csrutil status");
-        if (sipStatusOutput.exitCode === 0) {
-            if (sipStatusOutput.stdout.includes("System Integrity Protection status: enabled.")) {
-                return true;
-            }
-            if (sipStatusOutput.stdout.includes("System Integrity Protection status: disabled.")) {
-                return false;
-            }
-        }
-        return undefined;
-    }
-    catch (e) {
-        logger.warning(`Failed to determine if System Integrity Protection was enabled: ${e}`);
-        return undefined;
     }
 }
 function cleanupDatabaseClusterDirectory(config, logger,
