+++ lib/upload-sarif-action.js
@@ -34,7 +34,7 @@ const status_report_1 = require("./status-report");
 const upload_lib = __importStar(require("./upload-lib"));
 const util_1 = require("./util");
 async function sendSuccessStatusReport(startedAt, uploadStats, logger) {
-    const statusReportBase = await (0, status_report_1.createStatusReportBase)(status_report_1.ActionName.UploadSarif, "success", startedAt, undefined, await (0, util_1.checkDiskUsage)(), logger);
+    const statusReportBase = await (0, status_report_1.createStatusReportBase)(status_report_1.ActionName.UploadSarif, "success", startedAt, undefined, await (0, util_1.checkDiskUsage)(logger), logger);
     if (statusReportBase !== undefined) {
         const statusReport = {
             ...statusReportBase,
@@ -51,7 +51,7 @@ async function run() {
     (0, util_1.checkActionVersion)((0, actions_util_1.getActionVersion)(), gitHubVersion);
     const repositoryNwo = (0, repository_1.parseRepositoryNwo)((0, util_1.getRequiredEnvParam)("GITHUB_REPOSITORY"));
     const features = new feature_flags_1.Features(gitHubVersion, repositoryNwo, (0, actions_util_1.getTemporaryDirectory)(), logger);
-    const startingStatusReportBase = await (0, status_report_1.createStatusReportBase)(status_report_1.ActionName.UploadSarif, "starting", startedAt, undefined, await (0, util_1.checkDiskUsage)(), logger);
+    const startingStatusReportBase = await (0, status_report_1.createStatusReportBase)(status_report_1.ActionName.UploadSarif, "starting", startedAt, undefined, await (0, util_1.checkDiskUsage)(logger), logger);
     if (startingStatusReportBase !== undefined) {
         await (0, status_report_1.sendStatusReport)(startingStatusReportBase);
     }
@@ -74,7 +74,7 @@ async function run() {
             : (0, util_1.wrapError)(unwrappedError);
         const message = error.message;
         core.setFailed(message);
-        const errorStatusReportBase = await (0, status_report_1.createStatusReportBase)(status_report_1.ActionName.UploadSarif, (0, status_report_1.getActionsStatus)(error), startedAt, undefined, await (0, util_1.checkDiskUsage)(), logger, message, error.stack);
+        const errorStatusReportBase = await (0, status_report_1.createStatusReportBase)(status_report_1.ActionName.UploadSarif, (0, status_report_1.getActionsStatus)(error), startedAt, undefined, await (0, util_1.checkDiskUsage)(logger), logger, message, error.stack);
         if (errorStatusReportBase !== undefined) {
             await (0, status_report_1.sendStatusReport)(errorStatusReportBase);
         }
